apps/api/requirements.txt
Django>=5.0,<6
djangorestframework>=3.15,<4
psycopg[binary]>=3.1,<4
gunicorn>=22,<24
python-dotenv>=1.0,<2
